+++ UET/Redpoint.Uet.BuildPipeline/Executors/Engine/DefaultEngineWorkspaceProvider.cs
@@ -7,11 +7,11 @@
 
     internal class DefaultEngineWorkspaceProvider : IEngineWorkspaceProvider
     {
-        private readonly IDynamicWorkspaceProvider _workspaceProvider;
+        private readonly IWorkspaceProvider _workspaceProvider;
         private readonly IReservationManagerForUet _reservationManagerForUet;
 
         public DefaultEngineWorkspaceProvider(
-            IDynamicWorkspaceProvider workspaceProvider,
+            IWorkspaceProvider workspaceProvider,
             IReservationManagerForUet reservationManagerForUet)
         {
             _workspaceProvider = workspaceProvider;
@@ -25,25 +25,12 @@
         {
             if (buildEngineSpecification._enginePath != null)
             {
-                if (_workspaceProvider.ProvidesFastCopyOnWrite)
-                {
-                    return await _workspaceProvider.GetWorkspaceAsync(
-                        new FolderSnapshotWorkspaceDescriptor
-                        {
-                            SourcePath = buildEngineSpecification._enginePath,
-                            WorkspaceDisambiguators = new[] { workspaceSuffix },
-                        },
-                        cancellationToken).ConfigureAwait(false);
-                }
-                else
-                {
-                    return await _workspaceProvider.GetWorkspaceAsync(
-                        new FolderAliasWorkspaceDescriptor
-                        {
-                            AliasedPath = buildEngineSpecification._enginePath
-                        },
-                        cancellationToken).ConfigureAwait(false);
-                }
+                return await _workspaceProvider.GetWorkspaceAsync(
+                    new FolderAliasWorkspaceDescriptor
+                    {
+                        AliasedPath = buildEngineSpecification._enginePath
+                    },
+                    cancellationToken).ConfigureAwait(false);
             }
             else if (buildEngineSpecification._uefsPackageTag != null)
             {
