--- UET/Redpoint.Uet.Workspace/IDynamicWorkspaceProvider.cs
@@ -1,7 +1,0 @@
-﻿namespace Redpoint.Uet.Workspace
-{
-    public interface IDynamicWorkspaceProvider : IWorkspaceProviderBase
-    {
-        bool UseWorkspaceVirtualisation { get; set; }
-    }
-}
