--- UET/Redpoint.Uet.Workspace/IPhysicalWorkspaceProvider.cs
@@ -1,6 +1,0 @@
-﻿namespace Redpoint.Uet.Workspace
-{
-    public interface IPhysicalWorkspaceProvider : IWorkspaceProviderBase
-    {
-    }
-}
