--- UET/Redpoint.Uet.Workspace/IVirtualWorkspaceProvider.cs
@@ -1,6 +1,0 @@
-﻿namespace Redpoint.Uet.Workspace
-{
-    public interface IVirtualWorkspaceProvider : IWorkspaceProviderBase
-    {
-    }
-}
