+++ UET/Redpoint.Uet.Workspace/IWorkspaceProvider.cs
@@ -1,14 +1,6 @@
 ﻿namespace Redpoint.Uet.Workspace
 {
-    using Redpoint.Uet.Workspace.Descriptors;
-    using System.Threading.Tasks;
-
-    public interface IWorkspaceProviderBase
+    public interface IWorkspaceProvider : IWorkspaceProviderBase
     {
-        bool ProvidesFastCopyOnWrite { get; }
-
-        Task<IWorkspace> GetWorkspaceAsync(
-            IWorkspaceDescriptor workspaceDescriptor,
-            CancellationToken cancellationToken);
     }
 }
